# Changed attributes Ip, Ips, Body, Query, Params and Files to functions