Moved runtime parsing of request inputs to functions so they only execute on demand.