@ricallinson 2015-10-06
Current master bench:
BenchmarkOneMiddleware-4       	  500000	      2979 ns/op
BenchmarkFourMiddlewares-4     	  200000	      6101 ns/op
BenchmarkOneRoute-4            	  300000	      5666 ns/op
BenchmarkFourRoutes-4          	  200000	      6810 ns/op
BenchmarkFourRoutesMatchFirst-4	  200000	      5918 ns/op
BenchmarkFourRoutesNoMatch-4   	  200000	      5920 ns/op

This branch bench:
BenchmarkOneMiddleware-4       	  500000	      2279 ns/op
BenchmarkFourMiddlewares-4     	  300000	      5641 ns/op
BenchmarkOneRoute-4            	  300000	      4427 ns/op
BenchmarkFourRoutes-4          	  300000	      5650 ns/op
BenchmarkFourRoutesMatchFirst-4	  300000	      4716 ns/op
BenchmarkFourRoutesNoMatch-4   	  300000	      5145 ns/op